requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.13
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-extensions==3.1.5
django-heroku==0.3.1
django-render-block==0.9.1
django-templated-email==3.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
html2text==2020.1.16
idna==3.3
oauthlib==3.2.0
paystackapi==2.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.10
whitenoise==6.2.0
